Added auth.js,passport.js, implemented authorization for  api query

+++ index.js
@@ -10,6 +10,9 @@ const Users = Models.User;
 const app = express();
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
+let auth = require('./auth')(app);
+const passport = require('passport');
+require('./passport');
 //log to console
 app.use(morgan('common'));
 // Serve static files from the 'public' directory
@@ -20,7 +23,7 @@ app.get('/', (req, res) => {
   res.send('Welcome to myFlix!');
 });
 //return movies in json format
-app.get('/movies', async (req, res) => {
+app.get('/movies', passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.find()
     .then((movies) => {
       res.status(201).json(movies);
@@ -31,7 +34,7 @@ app.get('/movies', async (req, res) => {
     });
 });
 //Gets the data about a movie, by name
-app.get('/movies/:title', async (req, res) => {
+app.get('/movies/:title', passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.findOne({ Title: req.params.title })
     .then((movies) => {
       res.status(201).json(movies);
@@ -42,7 +45,7 @@ app.get('/movies/:title', async (req, res) => {
     });
 });
 // Gets the data about a genre (description) by name
-app.get('/movies/genre/:genreName', async (req, res) => {
+app.get('/movies/genre/:genreName', passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.findOne({ 'Genre.Name': req.params.genreName })
     .then((movies) => {
       res.status(201).send(movies.Genre.Description);
@@ -54,7 +57,7 @@ app.get('/movies/genre/:genreName', async (req, res) => {
 });
 
 // Gets the data about a director by name
-app.get('/movies/director/:name', async (req, res) => {
+app.get('/movies/director/:name', passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.findOne({ 'Director.Name': req.params.name })
     .then((movies) => {
       res.status(201).json(movies.Director);
@@ -88,7 +91,7 @@ app.post('/users', async (req, res) => {
   await Users.findOne({ UserName: req.body.UserName })
     .then((user) => {
       if (user) {
-        return res.status(400).send(req.body.UserName + 'already exists');
+        return res.status(400).send(req.body.UserName + ' already exists');
       } else {
         Users.create({
           UserName: req.body.UserName,
@@ -122,7 +125,10 @@ app.post('/users', async (req, res) => {
   (required)
   Birthday: Date
 }*/
-app.put('/users/:userName', async (req, res) => {
+app.put('/users/:userName', passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if (req.user.UserName !== req.params.userName) {
+    return res.status(400).send('Permission denied');
+  }
   await Users.findOneAndUpdate(
     { UserName: req.params.userName },
     {
@@ -145,35 +151,54 @@ app.put('/users/:userName', async (req, res) => {
 });
 
 // Allow users to add a movie to their list of favorites
-app.post('/users/:userName/movies/:movieId', async (req, res) => {
-  await Users.findOneAndUpdate(
-    { UserName: req.params.userName },
-    { $push: { FavoriteMovies: req.params.movieId } }
-  )
-    .then((user) => {
-      res.json(user);
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send('Error:' + err);
-    });
-});
+app.post(
+  '/users/:userName/movies/:movieId',
+  passport.authenticate('jwt', { session: false }),
+  async (req, res) => {
+    if (req.user.UserName !== req.params.userName) {
+      return res.status(400).send('Permission denied');
+    }
+    await Users.findOneAndUpdate(
+      { UserName: req.params.userName },
+      { $push: { FavoriteMovies: req.params.movieId } },
+      { new: true }
+    )
+
+      .then((user) => {
+        res.json(user);
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send('Error:' + err);
+      });
+  }
+);
 // Allow users to remove a movie from their list of favorites
-app.delete('/users/:userName/movies/:movieId', async(req, res) => {
-  await Users.findOneAndUpdate(
-    { UserName: req.params.userName },
-    { $pull: { FavoriteMovies: req.params.movieId } }
-  )
-    .then((user) => {
-      res.json(user);
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send('Error:' + err);
-    });
-});
+app.delete(
+  '/users/:userName/movies/:movieId',
+  passport.authenticate('jwt', { session: false }),
+  async (req, res) => {
+    if (req.user.UserName !== req.params.userName) {
+      return res.status(400).send('Permission denied');
+    }
+    await Users.findOneAndUpdate(
+      { UserName: req.params.userName },
+      { $pull: { FavoriteMovies: req.params.movieId } }
+    )
+      .then((user) => {
+        res.json(user);
+      })
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send('Error:' + err);
+      });
+  }
+);
 // Allow existing users to deregister
-app.delete('/users/:userName', async (req, res) => {
+app.delete('/users/:userName', passport.authenticate('jwt', { session: false }) ,async (req, res) => {
+  if (req.user.UserName !== req.params.userName) {
+    return res.status(400).send('Permission denied');
+  }
   await Users.findOneAndDelete({ UserName: req.params.userName })
     .then((user) => {
       if (!user) {
